Moved all types to separate folder - renamed
Added institution query to backend definition
