+++ matchmaker/query_engine/backends/optimised_scopus_meta/__init__.py
@@ -4,9 +4,9 @@ from matchmaker.query_engine.backends.scopus import PaperSearchQueryEngine as Sc
 from matchmaker.query_engine.backends.scopus import AuthorSearchQueryEngine as ScopusAuthorSearchQueryEngine
 from matchmaker.query_engine.backends.scopus import InstitutionSearchQueryEngine as ScopusInstitutionSearchQueryEngine
 
-from matchmaker.query_engine.data_types import AuthorData, PaperData, InstitutionData
-from matchmaker.query_engine.query_types import AuthorSearchQuery, PaperSearchQuery, InstitutionSearchQuery
-from matchmaker.query_engine.selector_types import AuthorDataAllSelected, PaperDataSelector, PaperDataAllSelected, AuthorDataSelector
+from matchmaker.query_engine.types.data import AuthorData, PaperData, InstitutionData
+from matchmaker.query_engine.types.query import AuthorSearchQuery, PaperSearchQuery, InstitutionSearchQuery
+from matchmaker.query_engine.types.selector import AuthorDataAllSelected, PaperDataSelector, PaperDataAllSelected, AuthorDataSelector
 from matchmaker.query_engine.slightly_less_abstract import AbstractNativeQuery
 from matchmaker.query_engine.slightly_less_abstract import SlightlyLessAbstractQueryEngine
 from matchmaker.query_engine.backend import Backend
